// twitter-clone-romanets-mongodb-docker-client/src/store/types.ts
export enum LoadingStatus {
  //данные уже были загружены или скачены
  LOADED = 'LOADED',   
  //данные загружаются или скачиваются
  LOADING = 'LOADING',  
  //данные загружены с ошибкой или скачены с ошибкой
  ERROR = 'ERROR',      
  //данные никогда не загружались или никогда не скачивались
  NEVER = 'NEVER',      
  //данные загружены успешно или скачены успешно
  SUCCESS = 'SUCCESS'   
}
  